Update index.js
I introduced the package react-native-webview because latest react-native package doesn't have webview anymore. I adjusted the view of the flutterwave web-page to fit the the webview screen. postMessage need to be well defined in the latest webview package, so i added the window.ReactNativeWebView. I enclosed the Webview component into a SafeAreaView to scale the app screen to contain the webview content.